requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==3.7.1
apturl==0.5.2
artifacts-keyring==0.3.3
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.0
azure-eventhub==5.11.5
bcrypt==3.2.0
beautifulsoup4==4.6.3
blinker==1.7.0
branca==0.7.0
Brlapi==0.8.3
Brotli==1.1.0
bs4==0.0.1
cachetools==5.3.3
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dataclasses-json==0.6.6
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
dj-database-url==2.1.0
Django==5.0.3
dnspython==2.1.0
duplicity==0.8.21
exceptiongroup==1.2.0
fasteners==0.14.1
Flask==3.0.1
folium==0.15.1
frozenlist==1.4.1
future==0.18.2
geographiclib==1.49
geopy==1.18.1
gitdb==4.0.11
GitPython==3.1.43
gpg==1.16.0
greenlet==3.0.3
gunicorn==21.2.0
gyp==0.1
h11==0.12.0
holehe==1.61
httpcore==0.13.7
httplib2==0.20.2
httpx==1.0.0b0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
jaraco.classes==3.3.0
jeepney==0.7.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kafka-python==1.4.6
keyring==23.13.1
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-text-splitters==0.0.1
langsmith==0.1.57
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.4
marshmallow==3.21.2
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.5
mypy-extensions==1.0.0
netifaces==0.11.0
numpy==1.24.4
oauthlib==3.2.0
olefile==0.46
orjson==3.10.3
outcome==1.3.0.post0
packaging==23.2
pandas==2.0.3
paramiko==2.9.3
passagem-taura==20231115.1
pexpect==4.8.0
Pillow==9.0.1
protobuf==4.25.3
ptyprocess==0.7.0
pyarrow==16.0.0
pycairo==1.20.1
pycups==2.0.1
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.9.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-editor==1.0.3
pytz==2023.3.post1
pyxdg==0.27
PyYAML==6.0
referencing==0.35.1
reportlab==3.6.8
requests==2.31.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.1
schedule==0.5.0
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.4.4
streamlit==1.34.0
style==1.1.0
systemd-python==234
tenacity==8.3.0
termcolor==2.4.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
trio==0.24.0
typing-inspect==0.9.0
typing_extensions==4.6.1
tzdata==2023.4
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
update==0.0.1
urllib3==2.1.0
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.6
watchdog==4.0.0
Werkzeug==3.0.1
whitenoise==6.6.0
xdg==5
xkit==0.0.0
XlsxWriter==3.0.8
xmltodict==0.11.0
xyzservices==2023.10.1
yarl==1.9.4
zipp==3.17.0
